Added post filtering and pizza creation features
In this commit, we have made several enhancements and modifications to both the BlogMvc and PizzaMvc projects.

In BlogMvc, we have added a feature to filter posts by title in `Index.cshtml` and handled exceptions in the `GetAllPosts` method in `PostManager.cs`.

In PizzaMvc, we have added a new action method `Create` in `HomeController.cs` to render the view for creating a new pizza. We have also added model validation in the `CreatePizza` method in `PizzaWebApiController.cs` and made the pizza name search case-insensitive in `GetAllPizzas` method and `GetPizzasByName` method in `PizzaManager.cs`.

We have updated the `Index.cshtml` and `Create.cshtml` views in PizzaMvc to support the new pizza creation and search functionalities.

Lastly, we have removed the `Create.cshtml` view from the up-to-date check and content included by default in `PizzaMvc.csproj`.